requirements.txt
yt-dlp
ffmpeg
requests
mutagen
python-dotenv
flask